Have discovered validators in ProtoData using a plain file
